requirements.txt
aiogram==3.10
python-dotenv==1.0.0
openai>=1.40.0
httpx>=0.27.0
PyMuPDF==1.24.5
python-pptx==0.6.23
groq==0.9.0
